Split test-soft-shutdown-pregel
