+++ tests/js/client/restart/test-soft-shutdown-cluster.js
@@ -28,60 +28,14 @@
 /// @author Copyright 2021, ArangoDB Inc, Cologne, Germany
 //////////////////////////////////////////////////////////////////////////////
 
-const _ = require('lodash');
 let jsunity = require('jsunity');
-const pu = require('@arangodb/testutils/process-utils');
-const request = require("@arangodb/request");
 const internal = require("internal");
 const db = require("internal").db;
 const time = internal.time;
 const wait = internal.wait;
-const statusExternal = internal.statusExternal;
 const {
   getCtrlCoordinators
 } = require('@arangodb/test-helper');
-
-var graph_module = require("@arangodb/general-graph");
-var EPS = 0.0001;
-let pregel = require("@arangodb/pregel");
-
-const graphName = "UnitTest_pregel";
-const vColl = "UnitTest_pregel_v", eColl = "UnitTest_pregel_e";
-
-function testAlgoStart(a, p) {
-  let pid = pregel.start(a, graphName, p);
-  assertTrue(typeof pid === "string");
-  return pid;
-}
-
-function testAlgoCheck(pid) {
-  let stats = pregel.status(pid);
-  console.warn("Pregel status:", stats);
-  return stats.state !== "running" && stats.state !== "storing";
-}
-
-function waitForAlive(timeout, baseurl, data) {
-  let res;
-  let all = Object.assign(data || {}, { method: "get", timeout: 1, url: baseurl + "/_api/version" });
-  const end = time() + timeout;
-  while (time() < end) {
-    res = request(all);
-    if (res.status === 200 || res.status === 401 || res.status === 403) {
-      break;
-    }
-    console.warn("waiting for server response from url " + baseurl);
-    wait(0.5);
-  }
-  return res.status;
-};
-
-function restartInstance(arangod) {
-  let options = _.clone(global.obj.options);
-  options.skipReconnect = false;
-  arangod.restartOneInstance();
-  waitForAlive(30, arangod.url, {});
-  arangod.checkArangoConnection(5);
-};
 
 function testSuite() {
   let cn = "UnitTestSoftShutdown";
@@ -100,12 +54,20 @@ function testSuite() {
 
       let coordinators = getCtrlCoordinators();
       assertTrue(coordinators.length > 0, "expect coordinators.length > 0");
+
+      // TODO: Pick a random coordinator?
       coordinator = coordinators[0];
     },
 
     tearDown : function() {
+      // Every test is supposed to initiate a soft shutdown
+      // hence, once we tear down the test we expect
+      // the coordinator to be gone
       coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
+      coordinator.restartOneInstance();
+
+      // we might have been connected to the coordinator that
+      // has been shut down
       db._drop(cn);
     },
 
@@ -370,141 +332,5 @@ function testSuite() {
   };
 }
 
-function testSuitePregel() {
-  'use strict';
-  let oldLogLevel;
-  let coordinator;
-  return {
-
-    /////////////////////////////////////////////////////////////////////////
-    /// @brief set up
-    /////////////////////////////////////////////////////////////////////////
-
-    setUpAll : function() {
-      oldLogLevel = arango.GET("/_admin/log/level").general;
-      arango.PUT("/_admin/log/level", { general: "info" });
-
-      let coordinators = getCtrlCoordinators();
-      assertTrue(coordinators.length > 0);
-      coordinator = coordinators[0];
-    },
-
-    tearDownAll : function () {
-      // And now it should shut down in due course...
-      coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
-
-      // restore previous log level for "general" topic;
-      arango.PUT("/_admin/log/level", { general: oldLogLevel });
-    },
-
-    setUp: function () {
-
-      var guck = graph_module._list();
-      var exists = guck.indexOf("demo") !== -1;
-      if (exists || db.demo_v) {
-        console.warn("Attention: graph_module gives:", guck, "or we have db.demo_v:", db.demo_v);
-        return;
-      }
-      var graph = graph_module._create(graphName);
-      db._create(vColl, { numberOfShards: 4 });
-      graph._addVertexCollection(vColl);
-      db._createEdgeCollection(eColl, {
-        numberOfShards: 4,
-        replicationFactor: 1,
-        shardKeys: ["vertex"],
-        distributeShardsLike: vColl
-      });
-
-      var rel = graph_module._relation(eColl, [vColl], [vColl]);
-      graph._extendEdgeDefinitions(rel);
-
-      var vertices = db[vColl];
-      var edges = db[eColl];
-
-
-      vertices.insert([{ _key: 'A', sssp: 3, pagerank: 0.027645934 },
-                       { _key: 'B', sssp: 2, pagerank: 0.3241496 },
-                       { _key: 'C', sssp: 3, pagerank: 0.289220 },
-                       { _key: 'D', sssp: 2, pagerank: 0.0329636 },
-                       { _key: 'E', sssp: 1, pagerank: 0.0682141 },
-                       { _key: 'F', sssp: 2, pagerank: 0.0329636 },
-                       { _key: 'G', sssp: -1, pagerank: 0.0136363 },
-                       { _key: 'H', sssp: -1, pagerank: 0.01363636 },
-                       { _key: 'I', sssp: -1, pagerank: 0.01363636 },
-                       { _key: 'J', sssp: -1, pagerank: 0.01363636 },
-                       { _key: 'K', sssp: 0, pagerank: 0.013636363 }]);
-
-      edges.insert([{ _from: vColl + '/B', _to: vColl + '/C', vertex: 'B' },
-                    { _from: vColl + '/C', _to: vColl + '/B', vertex: 'C' },
-                    { _from: vColl + '/D', _to: vColl + '/A', vertex: 'D' },
-                    { _from: vColl + '/D', _to: vColl + '/B', vertex: 'D' },
-                    { _from: vColl + '/E', _to: vColl + '/B', vertex: 'E' },
-                    { _from: vColl + '/E', _to: vColl + '/D', vertex: 'E' },
-                    { _from: vColl + '/E', _to: vColl + '/F', vertex: 'E' },
-                    { _from: vColl + '/F', _to: vColl + '/B', vertex: 'F' },
-                    { _from: vColl + '/F', _to: vColl + '/E', vertex: 'F' },
-                    { _from: vColl + '/G', _to: vColl + '/B', vertex: 'G' },
-                    { _from: vColl + '/G', _to: vColl + '/E', vertex: 'G' },
-                    { _from: vColl + '/H', _to: vColl + '/B', vertex: 'H' },
-                    { _from: vColl + '/H', _to: vColl + '/E', vertex: 'H' },
-                    { _from: vColl + '/I', _to: vColl + '/B', vertex: 'I' },
-                    { _from: vColl + '/I', _to: vColl + '/E', vertex: 'I' },
-                    { _from: vColl + '/J', _to: vColl + '/E', vertex: 'J' },
-                    { _from: vColl + '/K', _to: vColl + '/E', vertex: 'K' }]);
-    },
-
-    //////////////////////////////////////////////////////////////////////////
-    /// @brief tear down
-    //////////////////////////////////////////////////////////////////////////
-
-    tearDown: function () {
-      graph_module._drop(graphName, true);
-    },
-
-    testPageRank: function () {
-      // Start a pregel run:
-      let pid = testAlgoStart("pagerank", { threshold: EPS / 1000, resultField: "result", store: true });
-
-      console.warn("Started pregel run:", pid);
-
-      // Now use soft shutdown API to shut coordinator down:
-      let res = arango.DELETE("/_admin/shutdown?soft=true");
-      console.warn("Shutdown got:", res);
-      assertEqual("OK", res, `expecting res = ${res} == "OK"`);
-      let status = arango.GET("/_admin/shutdown");
-      assertTrue(status.softShutdownOngoing);
-      assertEqual(1, status.pregelConductors, `expecting status.pregelConductors = ${status.pregelConductors} == 1`);
-
-      // It should fail to create a new pregel run:
-      let gotException = false;
-      try {
-        testAlgoStart("pagerank", { threshold: EPS / 10, resultField: "result", store: true });
-      } catch(e) {
-        console.warn("Got exception for new run, good!", JSON.stringify(e));
-        gotException = true;
-      }
-      assertTrue(gotException, `expect to catch exception for new pregel run.`);
-
-      status = arango.GET("/_admin/shutdown");
-      assertTrue(status.softShutdownOngoing, `expect status.softShutdownOngoing == true`);
-
-      let startTime = time();
-      while (!testAlgoCheck(pid)) {
-        console.warn("Pregel still running...");
-        wait(0.2);
-        if (time() - startTime > 120) {
-          assertTrue(false, "Pregel did not finish in time.");
-        }
-      }
-
-      status = arango.GET("/_admin/shutdown");
-      assertTrue(status.softShutdownOngoing);
-    },
-
-  };
-}
-
-jsunity.run(testSuitePregel);
 jsunity.run(testSuite);
 return jsunity.done();
